extract demoworld to assembly
