+++ Explorer/Assets/DCL/Billboard/Billboard.asmdef
@@ -1,24 +1,25 @@
 {
-  "name": "Billboard",
-  "rootNamespace": "",
-  "references": [
-    "GUID:1b8e1e1bd01505f478f0369c04a4fb2f",
-    "GUID:4794e238ed0f65142a4aea5848b513e5",
-    "GUID:275e22790c04e9b47a5085d7b0c4432a",
-    "GUID:3c7b57a14671040bd8c549056adc04f5",
-    "GUID:0f4c0f120707fb74497f5d581b9858f8",
-    "GUID:fa7b3fdbb04d67549916da7bd2af58ab",
-    "GUID:08b2edf8f14cdd3408988fd7746b749c",
-    "GUID:d414ef88f3b15f746a4b97636b50dfb4",
-    "GUID:f51ebe6a0ceec4240a699833d6309b23"
-  ],
-  "includePlatforms": [],
-  "excludePlatforms": [],
-  "allowUnsafeCode": false,
-  "overrideReferences": false,
-  "precompiledReferences": [],
-  "autoReferenced": true,
-  "defineConstraints": [],
-  "versionDefines": [],
-  "noEngineReferences": false
+    "name": "Billboard",
+    "rootNamespace": "",
+    "references": [
+        "GUID:1b8e1e1bd01505f478f0369c04a4fb2f",
+        "GUID:4794e238ed0f65142a4aea5848b513e5",
+        "GUID:275e22790c04e9b47a5085d7b0c4432a",
+        "GUID:3c7b57a14671040bd8c549056adc04f5",
+        "GUID:0f4c0f120707fb74497f5d581b9858f8",
+        "GUID:fa7b3fdbb04d67549916da7bd2af58ab",
+        "GUID:08b2edf8f14cdd3408988fd7746b749c",
+        "GUID:d414ef88f3b15f746a4b97636b50dfb4",
+        "GUID:f51ebe6a0ceec4240a699833d6309b23",
+        "GUID:fd2b275602de4d638e01adafe118d5ea"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
 }
