--- Explorer/Assets/DCL/Billboard/Demo/World/DemoWorld.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: fafcf3f98d294ec0a3ec41ae0fb2892d
-timeCreated: 1702296519
