--- Explorer/Assets/DCL/Billboard/Demo/World/IDemoWorld.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 81b8fdec3a384272a42909d340403040
-timeCreated: 1702297360
