+++ Explorer/Assets/DCL/Billboard/Extensions.meta
@@ -1,3 +1,3 @@
 fileFormatVersion: 2
-guid: d5a9544c3ed64e009b87dacaa82c37b9
-timeCreated: 1702294995
+guid: 72dde118faf34caaa8127fad25376dbd
+timeCreated: 1702469337
