+++ Explorer/Assets/DCL/DemoWorlds/IDemoWorld.cs
@@ -1,7 +1,7 @@
 using Cysharp.Threading.Tasks;
 using System.Threading;
 
-namespace DCL.Billboard.Demo.World
+namespace DCL.DemoWorlds
 {
     public interface IDemoWorld
     {
