+++ Explorer/Assets/DCL/DemoWorlds/csc.rsp.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 533fea33316544009dd0a78b4e189e38
+timeCreated: 1702469080
